backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
backgroundremover==0.2.9
numpy<2.0.0
Pillow>=8.1.1,<10.0.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1

